requirements.txt
aiohttp==2.2.5
async-timeout==2.0.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
disco-py==0.0.12
docutils==0.14
gevent==1.2.2
greenlet==0.4.13
holster==2.0.0
humanize==0.5.1
idna==2.6
multidict==3.3.2
Pillow==5.0.0
psutil==5.4.3
Pygments==2.2.0
PyNaCl==1.2.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.44.0
websockets==3.4
Werkzeug==0.14.1
yarl==0.15.0
